ExhDetailInfo 공통부분 분리
