+++ src/components/regexh/ExhDetailInfoIntro.tsx
