+++ src/screens/exhibition/ExhDetailInfo.tsx
@@ -13,18 +13,7 @@ import {
   useFetchExhDetailInfo,
 } from '~/api/queries/exhibition';
 import LoadingModal from '~/components/common/modal/LoadingModal';
-import {dateToString} from '~/utils/date';
-import {
-  DEFAULT_TEXT,
-  LIGHT_GREY,
-  MAIN_COLOR,
-  MIDDLE_GREY,
-} from '~/components/common/colors';
-import {FONT_NAME} from '~/components/common/style';
-import {DEFAULT_IMAGE} from '@env';
-import ExhDetailInfoIntro from './detail/ExhDetailInfoIntro';
 import ExhReviewList from './detail/ExhReviewList';
-import ExhDetailHeart from './detail/ExhDetailHeart';
 import {TRenderEngineProvider} from 'react-native-render-html';
 import CustomTouchable from '~/components/common/CustomTouchable';
 import {
@@ -34,6 +23,7 @@ import {
 } from '~/components/common/icon';
 import ExhShare from './ExhShare';
 import {showToast} from '~/components/common/modal/toastConfig';
+import ExhDetailFormat from '~/components/regexh/ExhDetailFormat';
 
 type RootStackParamList = {
   ExhDetailInfo: {exhId: number};
@@ -68,6 +58,7 @@ const ExhDetailInfo: React.FC<Props> = ({route}) => {
   useEffect(() => {
     if (isFocused) {
       refetchDiaryList();
+      refetch();
     }
   }, [isFocused]);
 
@@ -136,23 +127,6 @@ const ExhDetailInfo: React.FC<Props> = ({route}) => {
     };
   }, [handlePressBack]);
 
-  const checkExhState = (): string => {
-    const currentDate = dateToString(new Date());
-
-    if (currentDate > data.exhPeriodEnd) {
-      // 진행상황 확인
-      return '종료';
-    } else if (
-      currentDate >= data.exhPeriodStart &&
-      currentDate <= data.exhPeriodEnd
-    ) {
-      return '진행중';
-    } else if (currentDate < data.exhPeriodStart) {
-      return '예정';
-    }
-    return '';
-  };
-
   const exhToHomepage = async (url: string) => {
     // 주어진 URL을 열 수 있는지 확인합니다.
     const supported = await Linking.canOpenURL(url);
@@ -174,6 +148,7 @@ const ExhDetailInfo: React.FC<Props> = ({route}) => {
         scrollEventThrottle={200}>
         {data && diaryData !== undefined && diaryData !== null && (
           <>
+            {/*전시회 세부정보 */}
             <TopLayer>
               <CustomTouchable onPress={handlePressBack}>
                 <BackButtonIcon />
@@ -199,51 +174,10 @@ const ExhDetailInfo: React.FC<Props> = ({route}) => {
                 )}
               </IconView>
             </TopLayer>
-            <TopView>
-              <BackgroundImage
-                source={{uri: `${data.poster ?? DEFAULT_IMAGE}`}}
-                blurRadius={40}
-                resizeMode="cover"
-                alt={'이미지 읽기 실패'}>
-                <ForegroundImage
-                  source={{uri: `${data.poster ?? DEFAULT_IMAGE}`}}
-                  resizeMode="contain"
-                  alt={'이미지 읽기 실패'}
-                />
-              </BackgroundImage>
-              <ExhDetailHeart exhId={exhId} hearState={data.favoriteExh} />
-            </TopView>
-            <InfoListView>
-              <Title>{data.exhName}</Title>
-              <StateView>
-                <StateText state={checkExhState()}>{checkExhState()}</StateText>
-              </StateView>
-              <InfoView>
-                <InfoTitle>{'장소'}</InfoTitle>
-                <Info>{data.gallery}</Info>
-              </InfoView>
-              <InfoView>
-                <InfoTitle>{'일정'}</InfoTitle>
-                <Info>{data.exhPeriodStart + ' ~ ' + data.exhPeriodEnd}</Info>
-              </InfoView>
-              <InfoView>
-                <InfoTitle>{'작가'}</InfoTitle>
-                {!data.painter ? (
-                  <Info>{'정보 없음'}</Info>
-                ) : (
-                  <Info>{data.painter}</Info>
-                )}
-              </InfoView>
-              <InfoView>
-                <InfoTitle>{'관람료'}</InfoTitle>
-                <Info>
-                  {data.fee}
-                  {'원'}
-                </Info>
-              </InfoView>
-            </InfoListView>
-            {/* 소개 */}
-            <ExhDetailInfoIntro intro={data.intro} />
+            {/**전시 상세정보 */}
+            <ExhDetailFormat data={data} state={'전시정보'} exhId={exhId} />
+
+            {/*전시 리뷰 */}
             <ExhReviewList exhId={exhId} diaryData={diaryData} />
           </>
         )}
@@ -262,13 +196,6 @@ const ContainerScroll = styled.ScrollView`
   background-color: white;
 `;
 
-const Title = styled.Text`
-  font-size: ${rf(19)}px;
-  color: ${DEFAULT_TEXT};
-  font-family: ${FONT_NAME};
-  line-height: ${wp(8)}px;
-`;
-
 const TopLayer = styled.View`
   flex-direction: row;
   justify-content: space-between; // 양 끝으로 버튼 배치
@@ -285,85 +212,4 @@ const IconView = styled.View`
   gap: ${wp(2)}px;
   justify-content: center;
   align-items: center;
-`;
-
-// poster section
-const TopView = styled.View`
-  flex: 1;
-  flex-direction: row;
-`;
-
-const BackgroundImage = styled.ImageBackground`
-  width: ${wp(100)}px;
-  height: ${hp(29)}px;
-`;
-
-const ForegroundImage = styled.Image`
-  width: 100%;
-  height: 100%;
-  align-items: center;
-`;
-
-// info section
-const InfoListView = styled.View`
-  flex-direction: column;
-  padding: ${wp(5.2)}px;
-  gap: ${wp(2)}px;
-  border-style: dashed;
-  border-bottom-width: ${wp(0.4)}px;
-  border-bottom-color: ${LIGHT_GREY};
-`;
-
-const StateView = styled.View`
-  flex-direction: row;
-  align-items: center;
-  padding-bottom: ${wp(2.9)}px;
-`;
-
-interface StateTextProps {
-  state: string;
-}
-
-const StateText = styled.Text<StateTextProps>`
-  font-size: ${rf(12)}px;
-  text-align: center;
-  color: ${(props: StateTextProps) =>
-    props.state === '진행중'
-      ? `${MAIN_COLOR}`
-      : props.state === '종료'
-        ? `${MIDDLE_GREY}`
-        : '#fee500'};
-  font-family: ${FONT_NAME};
-  padding-top: ${wp(1.1)}px;
-  padding-bottom: ${wp(0.6)}px;
-  padding-left: ${wp(1.9)}px;
-  padding-right: ${wp(1.9)}px;
-  border-color: ${(props: StateTextProps) =>
-    props.state === '진행중'
-      ? `${MAIN_COLOR}`
-      : props.state === '종료'
-        ? `${MIDDLE_GREY}`
-        : '#fee500'};
-  border-width: ${wp(0.3)}px;
-  border-radius: ${wp(50)}px;
-`;
-
-const InfoView = styled.View`
-  flex-direction: row;
-  gap: ${wp(5)}px;
-`;
-
-const InfoTitle = styled.Text`
-  font-size: ${rf(14.2)}px;
-  color: ${MIDDLE_GREY};
-  font-family: ${FONT_NAME};
-`;
-
-const Info = styled.Text`
-  flex-grow: 1;
-  flex-shrink: 1;
-  flex-basis: 0%;
-  font-size: ${rf(14.2)}px;
-  color: ${DEFAULT_TEXT};
-  font-family: ${FONT_NAME};
 `;
